Change the ydl_opts merging stage when task creating.
To keep the task's ydl_opts consistent even after alerting the global
ydl_opts, we change the chance to merge task's ydl_opts with global
ydl_opts to the time when task is created. In this way, the task's
ydl_opts is a snapshot of the current task is creating instead of
varing from time ot time.